Possible fix for loading breweries based on geolocation
Why: When fetch() tries to make an http request for "/api" it fails,
because "/api" is not a valid domain name (domain names are like
www.google.com). When fetch() fails it throws an exception that causes
the execution of getCords() to end before performing the fetch() for
breweries.

*---

How: When there are no exceptions, getCords() seems to work as intended.

*---